Seperate Schema and Models into seperate files for better isolation
