+++ database/dbModels.js
@@ -1,0 +1,26 @@
+const mongoose = require('mongoose'); // include DB object
+
+mongoose.connect('mongodb://localhost/restaurants');
+
+const availabilitySchema = mongoose.Schema({
+  id: {type:Number, unique: true},
+  availability: [{
+    day: Number,
+    hour: Number,
+    minute: Number
+  }],
+});
+
+const bookingsSchema = mongoose.Schema({
+  id: {type:Number, unique: true},
+  bookings: [{
+    day: Number,
+    bookings_count: Number
+  }],
+});
+
+let Availability = mongoose.model('Reservations', availabilitySchema);
+let Bookings = mongoose.model('Bookings', bookingsSchema);
+
+module.exports.availability = Availability;
+module.exports.bookings = Bookings;
